requirements.txt
fastapi>=0.115.8,<0.116.0
uvicorn>=0.34.0,<0.35.0
torch>=2.6.0,<3.0.0
torchvision>=0.21.0,<0.22.0
torchaudio>=2.6.0,<3.0.0
pillow>=11.1.0,<12.0.0
openai-clip>=1.0.1,<2.0.0
numpy>=2.2.2,<3.0.0
python-multipart>=0.0.20,<0.0.21
